+++ Library/Compiler.php
@@ -956,20 +956,25 @@ final class Compiler
      *
      * @throws Exception
      */
-    public function stubs($fromGenerate = false)
+    public function stubs(bool $fromGenerate = false)
     {
         if (!$fromGenerate) {
             $this->generate();
         }
 
         $this->logger->info('Generating stubs...');
-        $stubsGenerator = new Stubs\Generator($this->files, $this->config);
+
+        $stubsGenerator = new Stubs\Generator($this->files);
 
         $path = $this->config->get('path', 'stubs');
         $path = str_replace('%version%', $this->config->get('version'), $path);
         $path = str_replace('%namespace%', ucfirst($this->config->get('namespace')), $path);
 
-        $stubsGenerator->generate($path);
+        $stubsGenerator->generate(
+            $this->config->get('namespace'),
+            $path,
+            $this->config->get('indent', 'extra')
+        );
     }
 
     /**