+++ Library/Variable.php
@@ -842,9 +842,9 @@ class Variable implements TypeAwareInterface
             }
 
             $compilationContext->headersManager->add('kernel/memory');
+            $compilationContext->symbolTable->mustGrownStack(true);
 
             if (!$this->isLocalOnly()) {
-                $compilationContext->symbolTable->mustGrownStack(true);
                 if ($compilationContext->insideCycle) {
                     $this->mustInitNull = true;
                     $compilationContext->backend->initVar($this, $compilationContext, true, true);
@@ -863,9 +863,9 @@ class Variable implements TypeAwareInterface
             } else {
                 if ($this->variantInits > 0 || $compilationContext->insideCycle) {
                     $this->mustInitNull = true;
-                    $compilationContext->codePrinter->output('ZEPHIR_SINIT_NVAR('.$this->getName().');');
+                    $compilationContext->codePrinter->output('ZEPHIR_INIT_NVAR(&'.$this->getName().');');
                 } else {
-                    $compilationContext->codePrinter->output('ZEPHIR_SINIT_VAR('.$this->getName().');');
+                    $compilationContext->codePrinter->output('ZEPHIR_INIT_VAR(&'.$this->getName().');');
                 }
             }
 
@@ -938,20 +938,20 @@ class Variable implements TypeAwareInterface
             }
 
             $compilationContext->headersManager->add('kernel/memory');
+            $compilationContext->symbolTable->mustGrownStack(true);
             if (!$this->isLocalOnly()) {
-                $compilationContext->symbolTable->mustGrownStack(true);
                 if ($this->variantInits > 0 || $compilationContext->insideCycle) {
                     $this->mustInitNull = true;
-                    $compilationContext->codePrinter->output('ZEPHIR_INIT_LNVAR('.$this->getName().');');
+                    $compilationContext->codePrinter->output('ZEPHIR_INIT_NVAR(&'.$this->getName().');');
                 } else {
                     $compilationContext->backend->initVar($this, $compilationContext);
                 }
             } else {
                 if ($this->variantInits > 0 || $compilationContext->insideCycle) {
                     $this->mustInitNull = true;
-                    $compilationContext->codePrinter->output('ZEPHIR_SINIT_LNVAR('.$this->getName().');');
+                    $compilationContext->codePrinter->output('ZEPHIR_INIT_NVAR(&'.$this->getName().');');
                 } else {
-                    $compilationContext->codePrinter->output('ZEPHIR_SINIT_VAR('.$this->getName().');');
+                    $compilationContext->codePrinter->output('ZEPHIR_INIT_VAR(&'.$this->getName().');');
                 }
             }
             ++$this->variantInits;