+++ ext/test/exitdie.zep.h
@@ -16,6 +16,7 @@ ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_test_exitdie_testexit, 0, 0, IS_
 ZEND_BEGIN_ARG_INFO_EX(arginfo_test_exitdie_testexit, 0, 0, 0)
 #define arginfo_test_exitdie_testexit NULL
 #endif
+
 	ZEND_ARG_INFO(0, param)
 ZEND_END_ARG_INFO()
 
@@ -29,6 +30,7 @@ ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_test_exitdie_testdie, 0, 0, IS_V
 ZEND_BEGIN_ARG_INFO_EX(arginfo_test_exitdie_testdie, 0, 0, 0)
 #define arginfo_test_exitdie_testdie NULL
 #endif
+
 	ZEND_ARG_INFO(0, param)
 ZEND_END_ARG_INFO()