+++ ext/test/flow.zep.c
@@ -2136,19 +2136,21 @@ PHP_METHOD(Test_Flow, testFor39) {
 	long _1;
 	char _0;
 	zend_long i;
+	zephir_method_globals *ZEPHIR_METHOD_GLOBALS_PTR = NULL;
 	zval *this_ptr = getThis();
 
 	ZVAL_UNDEF(&_2);
 
+	ZEPHIR_MM_GROW();
 
 	i = 0;
-	ZEPHIR_SINIT_VAR(_2);
+	ZEPHIR_INIT_VAR(&_2);
 	ZVAL_STRING(&_2, "hello");
 	for (_1 = 0; _1 < Z_STRLEN_P(&_2); _1++) {
 		_0 = ZEPHIR_STRING_OFFSET(&_2, _1);
 		i++;
 	}
-	RETURN_LONG(i);
+	RETURN_MM_LONG(i);
 
 }