+++ ext/test/mcall.zep.c
@@ -159,7 +159,7 @@ PHP_METHOD(Test_Mcall, testCall3) {
 
 	ZEPHIR_MM_GROW();
 
-	ZEPHIR_RETURN_CALL_METHOD(this_ptr, "testmethod3", NULL, 45);
+	ZEPHIR_RETURN_CALL_METHOD(this_ptr, "testmethod3", NULL, 46);
 	zephir_check_call_status();
 	RETURN_MM();
 
@@ -222,7 +222,7 @@ PHP_METHOD(Test_Mcall, testCall6) {
 
 
 
-	ZEPHIR_RETURN_CALL_METHOD(this_ptr, "testmethod6", NULL, 46, a, b);
+	ZEPHIR_RETURN_CALL_METHOD(this_ptr, "testmethod6", NULL, 47, a, b);
 	zephir_check_call_status();
 	RETURN_MM();
 
@@ -299,7 +299,7 @@ PHP_METHOD(Test_Mcall, testCall9) {
 
 
 
-	ZEPHIR_CALL_METHOD(&_0, this_ptr, "testmethod6", NULL, 46, a, b);
+	ZEPHIR_CALL_METHOD(&_0, this_ptr, "testmethod6", NULL, 47, a, b);
 	zephir_check_call_status();
 	ZEPHIR_CALL_METHOD(&_1, this_ptr, "testmethod5", NULL, 0, a, b);
 	zephir_check_call_status();
@@ -347,7 +347,7 @@ PHP_METHOD(Test_Mcall, testCall12) {
 
 	ZEPHIR_MM_GROW();
 
-	ZEPHIR_RETURN_CALL_METHOD(this_ptr, "testmethod3", NULL, 45);
+	ZEPHIR_RETURN_CALL_METHOD(this_ptr, "testmethod3", NULL, 46);
 	zephir_check_call_status();
 	RETURN_MM();
 
@@ -410,7 +410,7 @@ PHP_METHOD(Test_Mcall, testCall15) {
 
 
 
-	ZEPHIR_RETURN_CALL_METHOD(this_ptr, "testmethod6", NULL, 46, a, b);
+	ZEPHIR_RETURN_CALL_METHOD(this_ptr, "testmethod6", NULL, 47, a, b);
 	zephir_check_call_status();
 	RETURN_MM();
 
@@ -467,7 +467,7 @@ PHP_METHOD(Test_Mcall, testCall18) {
 
 	ZEPHIR_MM_GROW();
 
-	ZEPHIR_RETURN_CALL_METHOD(this_ptr, "testmethod7", NULL, 47);
+	ZEPHIR_RETURN_CALL_METHOD(this_ptr, "testmethod7", NULL, 48);
 	zephir_check_call_status();
 	RETURN_MM();
 
@@ -584,7 +584,7 @@ PHP_METHOD(Test_Mcall, testCall22) {
 				_0 = 1;
 			}
 			i = _1;
-			ZEPHIR_CALL_METHOD(&_3$$3, this_ptr, "testmethod21", &_4, 48, p, p);
+			ZEPHIR_CALL_METHOD(&_3$$3, this_ptr, "testmethod21", &_4, 49, p, p);
 			zephir_check_call_status();
 			j += zephir_get_numberval(&_3$$3);
 		}
@@ -930,7 +930,7 @@ PHP_METHOD(Test_Mcall, testCallablePass) {
 	ZEPHIR_INIT_VAR(&_0);
 	ZEPHIR_INIT_NVAR(&_0);
 	zephir_create_closure_ex(&_0, NULL, test_11__closure_ce, SL("__invoke"));
-	ZEPHIR_RETURN_CALL_METHOD(&a, "setcallable", NULL, 49, &_0);
+	ZEPHIR_RETURN_CALL_METHOD(&a, "setcallable", NULL, 50, &_0);
 	zephir_check_call_status();
 	RETURN_MM();
 
@@ -962,7 +962,7 @@ PHP_METHOD(Test_Mcall, testCallableArrayThisMethodPass) {
 	ZEPHIR_INIT_VAR(&_1);
 	ZVAL_STRING(&_1, "bb");
 	zephir_array_fast_append(&_0, &_1);
-	ZEPHIR_RETURN_CALL_METHOD(&a, "setcallable", NULL, 49, &_0);
+	ZEPHIR_RETURN_CALL_METHOD(&a, "setcallable", NULL, 50, &_0);
 	zephir_check_call_status();
 	RETURN_MM();
 
@@ -985,7 +985,7 @@ PHP_METHOD(Test_Mcall, aa) {
 		ZEPHIR_CALL_METHOD(NULL, &a, "__construct", NULL, 0);
 		zephir_check_call_status();
 	}
-	ZEPHIR_RETURN_CALL_METHOD(&a, "bb", NULL, 50);
+	ZEPHIR_RETURN_CALL_METHOD(&a, "bb", NULL, 51);
 	zephir_check_call_status();
 	RETURN_MM();