+++ ext/test/regexdna.zep.c
@@ -179,7 +179,7 @@ PHP_METHOD(Test_RegexDNA, process) {
 	ZEPHIR_CONCAT_SVS(&_1, "/", &stuffToRemove, "/mS");
 	ZEPHIR_INIT_NVAR(&_0);
 	ZVAL_STRING(&_0, "");
-	ZEPHIR_CALL_FUNCTION(&_2, "preg_replace", &_3, 71, &_1, &_0, &contents);
+	ZEPHIR_CALL_FUNCTION(&_2, "preg_replace", &_3, 72, &_1, &_0, &contents);
 	zephir_check_call_status();
 	ZEPHIR_CPY_WRT(&contents, &_2);
 	ZEPHIR_INIT_VAR(&codeLength);
@@ -192,7 +192,7 @@ PHP_METHOD(Test_RegexDNA, process) {
 			ZVAL_COPY(&regex, _4);
 			zend_print_zval(&regex, 0);
 			php_printf("%s", " ");
-			ZEPHIR_INIT_LNVAR(_6$$3);
+			ZEPHIR_INIT_NVAR(&_6$$3);
 			ZEPHIR_CONCAT_SVS(&_6$$3, "/", &regex, "/iS");
 			ZEPHIR_INIT_NVAR(&_7$$3);
 			zephir_preg_match(&_7$$3, &_6$$3, &contents, &discard, 1, 0 , 0  TSRMLS_CC);
@@ -212,7 +212,7 @@ PHP_METHOD(Test_RegexDNA, process) {
 			zephir_check_call_status();
 				zend_print_zval(&regex, 0);
 				php_printf("%s", " ");
-				ZEPHIR_INIT_LNVAR(_8$$4);
+				ZEPHIR_INIT_NVAR(&_8$$4);
 				ZEPHIR_CONCAT_SVS(&_8$$4, "/", &regex, "/iS");
 				ZEPHIR_INIT_NVAR(&_9$$4);
 				zephir_preg_match(&_9$$4, &_8$$4, &contents, &discard, 1, 0 , 0  TSRMLS_CC);
@@ -223,7 +223,7 @@ PHP_METHOD(Test_RegexDNA, process) {
 		}
 	}
 	ZEPHIR_INIT_NVAR(&regex);
-	ZEPHIR_CALL_FUNCTION(&_2, "preg_replace", &_3, 71, &vIUB, &vIUBnew, &contents);
+	ZEPHIR_CALL_FUNCTION(&_2, "preg_replace", &_3, 72, &vIUB, &vIUBnew, &contents);
 	zephir_check_call_status();
 	ZEPHIR_CPY_WRT(&contents, &_2);
 	php_printf("%c", '\n');