+++ unit-tests/fixtures/stubs/issue1922/expected/Test.zep.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Stubs;
+
+
+class Test
+{
+
+    /**
+     * @param string $key
+     * @return mixed|null
+     */
+    public function getVar(string $key) {}
+
+}